Create two functions for album data
function get_albums uses spotipy to get a list of songs from an album
using the album url.

get_all_albums uses get_albums to get a list of songs for each of
mitski and phoebe's albums. it takes no arguments, and returns a list
of song names for each album